requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
psycopg2-binary
gunicorn  # For production
python-dotenv
